combined gift and request menus into dynamic 'ContributionMenu'
